Fix GraphViz icon to work with Rust build
The Rust build uses usvg[0] as part of the rendering process. usvg does
not support <circle> so we have to convert it over to <path> instead.

Interestingly the docs say usvg doesn't support <rect> but it seems to
work fine in practice. That might be something to convert to <path> as
well just to be safe in the future.

[0] https://github.com/RazrFalcon/resvg/tree/master/usvg